@@ -1,74 +1,125 @@
 [package]
-name = "code2prompt"
-version = "3.0.2" # You can update this later
-authors = ["Mufeed VH <mufeed@lyminal.space>","Olivier D'Ancona <olivier.dancona@master.hes-so.ch>"]
+name        = "code2prompt-tui"
+version     = "1.0.0"
+authors     = [
+                "Tahsin Kocaman",
+                "Mufeed VH <mufeed@lyminal.space>",
+                "Olivier D'Ancona <olivier.dancona@master.hes-so.ch>",
+            ]
 description = "A command-line (CLI) tool to generate an LLM prompt from codebases of any size, fast."
-keywords = ["code", "prompt", "llm", "gpt", "ai"]
-categories = ["command-line-utilities", "development-tools"]
-homepage = "https://github.com/mufeedvh/code2prompt"
+keywords    = ["code", "prompt", "llm", "gpt", "ai"]
+categories  = ["command-line-utilities", "development-tools"]
+homepage    = "https://github.com/mufeedvh/code2prompt"
 documentation = "https://github.com/mufeedvh/code2prompt"
-repository = "https://github.com/mufeedvh/code2prompt"
-license = "MIT"
-readme = "README.md"
-exclude = [".github/*", ".assets/*"]
-edition = "2021" # Sticking to 2021 for now for wider compatibility
+repository  = "https://github.com/mufeedvh/code2prompt"
+license     = "MIT"
+readme      = "README.md"
 
-[lib]
-name = "code2prompt"
-path = "src/lib.rs"
-
-[[bin]]
-name = "code2prompt"
-path = "src/main.rs"
+edition     = "2024"
+resolver    = "2"
 
 [dependencies]
-# Core Engine Dependencies
+# --- Core: always compiled ---------------------------------------------------
 anyhow          = "1.0"
-clap            = { version = "4.5", features = ["derive"] }
-dashmap         = "6.1"
-derive_builder  = "0.20.2"
+crossbeam-channel   = "0.5.15"
+derive_builder  = "0.20"
+dirs            = "6.0.0"
 glob            = "0.3"
 globset         = "0.4"
 handlebars      = { version = "6.2", default-features = false }
+hex             = "0.4.3"
 ignore          = "0.4"
-log             = "0.4"
 once_cell       = "1.20"
 regex           = { version = "1.11", default-features = false, features = ["std", "unicode-perl"] }
+rusqlite        = { version = "0.37", features = ["bundled", "serde_json"] }
 serde           = { version = "1.0", features = ["derive"] }
-serde_json      = "1.0"
-termtree        = "0.5"
-tiktoken-rs     = "0.7.0"
+serde_json      = "1.0.141"
+
+# --- Tokenisation -----------------------------------------------------------
+dashmap         = { version = "6.1", optional = true }
+rustc-hash      = { version = "2.1.1", optional = true }
+tiktoken-rs     = { version = "0.7.0", optional = true }
 
 # UI / Optional Dependencies
 atty            = { version = "0.2", optional = true }
+env_logger      = { version = "0.11.8", optional = true }
+terminal_size   = { version = "0.4", optional = true }
+unicode-width   = { version = "0.2", optional = true }
+flate2          = { version = "1.1", optional = true }
+rayon           = { version = "1.10", optional = true }
+
+# --- Optional colours --------------------------------------------------------
+colored         = { version = "3.0", optional = true }
+lscolors        = { version = "0.20", optional = true, default-features = false }
+
+# --- Optional interactive CLI -----------------------------------------------
+inquire         = { version = "0.7", optional = true, default-features = false, features = ["crossterm"] }
+indicatif       = { version = "0.18", default-features = false, features = ["improved_unicode"], optional = true }
+
+# --- Clipboard (Linux, macOS, Windows) --------------------------------------
 arboard         = { version = "3.4", default-features = false, optional = true }
-colored         = { version = "2.0", optional = true } # Note: version 2 for wider compat
-lscolors        = { version = "0.20", default-features = false, optional = true }
+
+# --- Git support ------------------------------------------------------------
 git2            = { version = "0.19", default-features = false, features = ["vendored-libgit2"], optional = true }
-indicatif       = { version = "0.17.0", default-features = false, features = ["improved_unicode"], optional = true } # Note: version 0.17
-inquire         = { version = "0.7", optional = true }
-env_logger      = { version = "0.11", optional = true }
-terminal_size   = { version = "0.3.0", optional = true } # Note: version 0.3
-unicode-width   = { version = "0.2.1", optional = true } # Note: version 0.1
+
+# --- Tree printing ----------------------------------------------------------
+termtree        = "0.5"
+
+# --- TUI dependencies -------------------------------------------------------
+ratatui         = { version = "0.29", optional = true, default-features = false, features = ["crossterm"] }
+crossterm       = { version = "0.29", optional = true }
+
+# --- Other helpers ----------------------------------------------------------
+clap            = { version = "4.5", features = ["derive", "wrap_help"] }
+log             = { version = "0.4", optional = true }
+sha2            = "0.10"
+confy           = "1.0.0"
+toml            = "0.9.2"
+thousands       = "0.2.0"
+bitflags        = "2.9.1"
 
 [features]
-default = ["colors", "clipboard", "git", "interactive", "logging"]
+default      = ["colors"]
 
-# Granular switches for optional functionality
-colors      = ["colored", "lscolors", "atty"]
-clipboard   = ["arboard"]
-git         = ["git2"]
-interactive = ["inquire", "indicatif", "terminal_size", "unicode-width", "atty"]
-logging     = ["env_logger"]
+# ───── Layered feature flags ───────────────────────────────────────────────
+cache        = ["dep:dashmap", "dep:rustc-hash", "dep:flate2", "dep:rayon"]
+clipboard    = ["dep:arboard"]
+colors       = ["dep:colored", "dep:lscolors"]
+git          = ["dep:git2"]
+interactive  = ["dep:inquire", "dep:indicatif"]
+logging      = ["dep:log", "dep:env_logger"]
+token_map    = ["dep:tiktoken-rs", "dep:atty", "dep:terminal_size", "dep:unicode-width"]
+tui          = ["dep:ratatui", "dep:crossterm", "dep:rustc-hash", "cache"]
+
+# Convenience “mega” feature
+full         = ["colors", "logging", "git", "tui", "interactive", "clipboard", "token_map"]
 
 [dev-dependencies]
-assert_cmd = "2.0"
-predicates = "3.1"
-serial_test = "3.2.0"
-tempfile = "3.10" # Note: version 3.10
+assert_cmd      = "2.0"
+filetime        = "0.2"
+predicates      = "3.1"
+quickcheck      = "1.0.3"
+quickcheck_macros = "1.1.0"
+serial_test     = "3.2.0"
+tempfile        = "3.10"
+
+[[bin]]
+name            = "code2prompt-tui"
+path            = "src/main.rs"
 
 [profile.release]
-lto = "thin"
-opt-level = 3
-panic = 'abort'
-codegen-units = 1
+opt-level       = "z"
+lto             = "thin"
+codegen-units   = 1
+debug           = 0
+strip           = "symbols"
+panic           = 'abort'
+
+[alias]
+full-build      = "build --release --features full"
+full-build-win  = "build --release --features full --target x86_64-pc-windows-gnu"
+full-run        = "run --release --features full"
+lint            = "clippy --all-targets --all-features -- -D warnings"
+check-full      = "check --all-targets --release --features full"
+
+exclude         = ["/.github/*", "/tests/data/*", "target/*"]
