@@ -1,5 +1,21 @@
-// src/ui/mod.rs
+pub mod cache;
 pub mod cli;
 pub mod clipboard;
+pub mod config;
+pub mod output;
+
+#[cfg(feature = "tui")]
+pub mod pane;
+
 pub mod template;
 pub mod token_map_view;
+pub mod tree_view;
+
+#[cfg(feature = "tui")]
+pub mod tree_arena;
+
+#[cfg(feature = "tui")]
+pub mod tree_pane;
+
+#[cfg(feature = "tui")]
+pub mod tui_select;
