@@ -3,7 +3,7 @@ use colored::*;
 use log::{debug, info};
 use predicates::prelude::*;
 use predicates::str::contains;
-use std::fs::{self, read_to_string, File};
+use std::fs::{self, File, read_to_string};
 use std::io::Write;
 use std::path::Path;
 use std::sync::Once;
