@@ -1,4 +1,6 @@
-use code2prompt::ui::template::{extract_undefined_variables, handlebars_setup, render_template};
+use code2prompt_tui::ui::template::{
+    extract_undefined_variables, handlebars_setup, render_template,
+};
 
 #[cfg(test)]
 mod tests {
@@ -52,10 +54,10 @@ mod tests {
         // It should probably find `user` and `goal`. The regex needs to be improved.
         // Let's test the current behavior.
         assert_eq!(variables, vec!["goal"]); // This is the actual current behavior.
-                                             // The regex is `[a-zA-Z_][a-zA-Z_0-9]*`,
-                                             // so it won't match the `#if user` part.
-                                             // It also won't match `user.name`.
-                                             // This test now correctly documents the existing limitations.
+        // The regex is `[a-zA-Z_][a-zA-Z_0-9]*`,
+        // so it won't match the `#if user` part.
+        // It also won't match `user.name`.
+        // This test now correctly documents the existing limitations.
     }
 
     #[test]
@@ -70,5 +72,37 @@ mod tests {
             Ok(output) => assert_eq!(output, "Hello, Bernard!"),
             Err(e) => panic!("Template rendering failed: {}", e),
         }
+    }
+
+    #[test]
+    fn test_extract_placeholders_simple() {
+        let template_str = "{{greeting}}, {{name}}! Your project is {{project_name}}.";
+        let mut variables = extract_placeholders(template_str).unwrap();
+        variables.sort();
+        assert_eq!(variables, vec!["greeting", "name", "project_name"]);
+    }
+
+    #[test]
+    fn test_extract_placeholders_ignores_builtins_and_paths() {
+        let template_str = r#"
+            Project: {{absolute_code_path}}
+            {{#each files}}
+                Path: {{this.path}}
+            {{/each}}
+            User var: {{goal}}
+            Diff: {{git_diff}}
+        "#;
+        let variables = extract_placeholders(template_str).unwrap();
+        // Should only find 'goal', as the others are built-in or complex paths.
+        assert_eq!(variables, vec!["goal"]);
+    }
+
+    #[test]
+    fn test_extract_placeholders_handles_duplicates() {
+        let template_str = "{{greeting}}, {{name}}! How is {{name}} today?";
+        let mut variables = extract_placeholders(template_str).unwrap();
+        variables.sort();
+        // `name` should only appear once in the output.
+        assert_eq!(variables, vec!["greeting", "name"]);
     }
 }
